refactoring the whole Application. Deleting index.html, renaming and improving main.html update LinkController.java to this changes update LinkControllerTest.java to this changes
