+++ src/main/java/de/hhu/propra/link/security/SecurityConfiguration.java
@@ -1,4 +1,4 @@
-package de.hhu.propra.link;
+package de.hhu.propra.link.security;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
