--- src/main/resources/templates/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<head>
-  <meta charset="UTF-8">
-  <title>Shorty</title>
-  <link href="../static/node_modules/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet"
-        th:href="@{node_modules/bootstrap/dist/css/bootstrap.min.css}">
-</head>
-<body>
-<div class="container pt-3">
-  <h1 class="text-center">
-    Shorty: URL Shortener
-  </h1>
-  <p class="lead text-center">
-    Create personalized short-links
-  </p>
-
-  <div class="text-center m-5">
-    <p>
-      You need to be an admin to create new short links.
-    </p>
-    <a class="btn btn-primary" th:href="@{/admin}">Admin Panel</a>
-  </div>
-</div>
-</body>
-</html>
